requirements.txt
loguru==0.7.2
web3==6.19.0
python-dotenv==1.1.0